load_cell_module/validation/requirements.txt
pyserial
sounddevice
numpy
